chromasonic/requirements.txt
# Core ML and Data Science
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
matplotlib>=3.4.0
seaborn>=0.11.0

# Image Processing
opencv-python>=4.5.0
Pillow>=8.3.0
scikit-image>=0.19.0

# Color Analysis
colorthief>=0.2.1
colorspacious>=1.1.2

# Audio Processing and Synthesis
librosa>=0.9.0
soundfile>=0.10.0
mido>=1.2.0
music21>=7.0.0
pyaudio>=0.2.11
pydub>=0.25.0

# Deep Learning
torch>=1.10.0
tensorflow>=2.8.0
transformers>=4.15.0

# Web Interface
flask>=2.0.0
flask-cors>=3.0.0
flask-socketio>=5.0.0
gunicorn>=20.1.0

# Utilities
tqdm>=4.62.0
joblib>=1.1.0
scipy>=1.7.0